+++ iti-react/Changelog.md
@@ -444,6 +444,9 @@ const myTab: Tab = ['tabName', 'Tab Label', { className: 'my-tab' }]
 # 1.9.2
 
 -   \*\*\* onChildReady:
-    -   Deprecate the old onChildReady - Add onChildReady2. See OnChildReady.ts for the migration path
+    -   Deprecate the old onChildReady
+    -   Add onChildReady2. See OnChildReady.ts for the migration path
+-   PaginationUtil:
+    -   Bring in resetPageIfFiltersChanged - Bring in an improved version preventNonExistentPage
 -   Core changes
 -   Add `validated-input` class to the `<div>` that wraps ValidatedInput, .etc